Improve testability of BrowserWindow mocks
So this is my first attempt at improving the way BrowserWindow is tested. The only way to create a mocked BrowserWindow now is to call the constructor, which makes sure that the array of windows that’s being tracked internally always represents all the windows that were actually constructed.

I added `onNewWindow` so tests can register a handler that is called whenever a new window is constructed. This way, tests can still mock certain functions / do whatever they want with the window.

Added `acceptFirstMouse` tests for bonus points.