Window nesting
I took a stab at nesting of window elements, which turend out to be quite hard to do.

Decided not to use `setParentWindow`, because of limitations on Windows: https://github.com/electron/electron/blob/master/docs/api/browser-window.md#platform-notices

This does mean that when nested, windows need to be constructed top down, instead of bottom up, because child windows need the reference of their parent in the constructor. This is not how fiber likes to do things.

As a side effect, doing this meant that windows might not be instantiated in their own `finalizeBeforeMount ` phase, which means that adding event handlers must wait untill `commitMount`.

I don’t think I have a good enough understanding of fiber to oversee the consequences of moving the event handlers, so any ideas/feedback/comments are very welcome.